requirements.txt
streamlit>=1.35.0
openai-whisper>=20231117
anthropic>=0.25.0
openpyxl>=3.1.2
psycopg2-binary>=2.9.9
numpy>=1.24.0
scipy>=1.11.0
soundfile>=0.12.1
